+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Basic macOS support. Currently untested.
 - Changed several key and gamepad functions to use `u32`, making it more ergonomic with key/gamepad constants.
+- Fixed unnecessary `&mut` in `load_image_ex` and `draw_poly_ex`.
 
 ## 0.9.1